initial game_options finished

+++ src/card.rs
@@ -13,7 +13,7 @@ pub enum Suit {
 
 pub type Id = u8;
 pub type Rank = u8;
-pub type Points = usize;
+pub type Points = i16;
 
 #[derive(Clone, Debug)]
 pub struct Card {
